Adding RMD File
This is the RMD file for the model.  This is where we will continuously work on changes.  After changes are implemented in the RMD file, we will knit the file into HMTL format and upload that separately.  At this moment, the linear regressions still need to be added to the model.